added tests for authentication
