+++ tests/test_registration.py
@@ -3,25 +3,26 @@ import unittest
 import requests
 from sqlalchemy import select
 
+from tests.mixins import DatabaseWiperTestMixin, RestAPITestMixin
 from tests.models import Application
-from tests.utils import DatabaseWiperTestMixin, RestAPITestUtilities, HTTP_METHODS, BASE_URL
+from tests.utils import HTTP_METHODS, BASE_URL
 
 
-class TestRegistration(DatabaseWiperTestMixin, RestAPITestUtilities, unittest.TestCase):
+class TestRegistration(DatabaseWiperTestMixin, RestAPITestMixin, unittest.TestCase):
     url = BASE_URL + "/register"
     invalid_methods = HTTP_METHODS - {"post"}
     required_fields = {"name", "password"}
 
-    user = dict(name="goatsy", password="goat")
+    application = dict(name="goatsy", password="goat")
 
     def test_can_register(self):
-        r = requests.post(self.url, json=self.user)
+        r = requests.post(self.url, json=self.application)
         self.assertEqual(r.status_code, requests.codes.created)
         self.assertIn("Authorization", r.headers.keys())
 
     def test_cannot_register_twice(self):
-        requests.post(self.url, json=self.user)
-        r = requests.post(self.url, json=self.user)
+        requests.post(self.url, json=self.application)
+        r = requests.post(self.url, json=self.application)
 
         self.assertEqual(r.status_code, requests.codes.conflict)
         self.assertIsNotNone(
@@ -31,7 +32,7 @@ class TestRegistration(DatabaseWiperTestMixin, RestAPITestUtilities, unittest.Te
         self.assertListEqual(r.json()["name"].keys(), ["message", "code"])
 
     def test_password_is_hashed(self):
-        r = requests.post(self.url, json=self.user)
+        r = requests.post(self.url, json=self.application)
         self.assertEqual(
             r.status_code,
             requests.codes.created,
@@ -39,9 +40,10 @@ class TestRegistration(DatabaseWiperTestMixin, RestAPITestUtilities, unittest.Te
         )
 
         result = self.database_connection.execute(
-            select([Application]).where(Application.name == self.user["name"])
+            select([Application]).where(Application.name == self.application["name"])
         ).first()
-        self.assertNotEqual(self.user["password"], result["password"], msg="Password is not hashed")
+        self.assertNotEqual(self.application["password"], result["password"], msg="Password is not hashed")
+
 
 if __name__ == '__main__':
     unittest.main()
